Use the scaled and cached image on candidate pages
The site was previously downloading the full sized image on candidate
pages; not only will this be slow and use a lot of bandwidth, currently
those images aren't cached properly by Varnish, so put unnecessary load
on the server.  This commit changes the candidate pages to use a
thumbnail generated by the image proxy.